Update ColorCode for red or gray tiles
Add ability to send a code (-1) for red tiles, indicating player has tapped FillInTheBlanks, and -2 for gray tiles, for words that are in the current list, but inactive (increasing playability by automatically filling in words that the player has already found a number of times). Moved the Colors struct to its own file, so that Colors can be accessed from the View Controller class.