Calculate inactive words, and display as gray
To increase playability and fun (by automatically finding, and displaying in gray, a percentage of frequently found words). Set some, but not all, of the mastered words to inactive when a new word list is generated. Display inactive words in gray, and do not count them as being either played, or found.