Update word list progress and display
Calculate number of inactive words, and subtract those from the length of the word list. Player is notified when word list is completed, now taking into account that some words may be inactive. When an inactive word is played by the player, the player is notified that this word has already been mastered.