Format Progress view table
The first of a series of tables. Each contains found words that have been persisted in Core Data. Formatted the table with a custom table cell so that each word is small, but readable, and uses contraints on a custom label to bring the margin close to the edge of the table. Each table will be one bar in a bar chart, its height indicating how many words have been found at each level. The bars will be in proportion to each other, and if there are too many words to fit, the table/bar can be scrolled to view all of them.