Pass data from Progress view to Level views
Create a level Int var, and wordsInLevel array of Word objects, in the level table view controller. In the Progress view, pass the data from the core data fetch into each level view.
Also, change constraints for level container views to set a narrow width, so that each will be a bar in a graph.