Format the graph level bars
Set the cell color of the table cells which make up the graph based on the level number. Create new outline and shadow images for the level bars and slice for scaling. Add a mask to the container view holding the level bars so that the corners can be rounded.