Fix 2 warnings of deprecated syntax in Swift 3
Fix the following two warnings for code that works in Swift 2 but will be deprecated in Swift 3:
WordListCell.swift:47:80: 'init(start:end:)' is deprecated: it will be removed in Swift 3.  Use the '..<' operator.
AtoZModel.swift:237:23: 'var' parameters are deprecated and will be removed in Swift 3
In this 2nd case, simply remove var from the parameter, and at the start of the function, add a var and set it equal to the parameter.