Add slight delay to fix swap tiles going to wrong position
Swap tiles should, when a tile is tapped, send a tile that's in a lower position, to an upper position, if one is open. Added a delay function, and called a slight delay in findVacancy(). Fixes a lot of the problems, but not perfectly. May affect the speedy feel of the tile interaction. To do: find a better solution.