Add completion handler to saveContext
Add a version of saveContext with a completion handler. In swapTile, use this version, and run the last 2 lines of swapTile contingent on success of saveContext. In preparation for ensuring that the context is saved before the next call to findVacancy. Attempting to fix the bug where tiles don't go to correct location when they are tapped in quick succession.