Fix bug where tapped tiles went to the wrong place
The tiles would sometimes, when tapped in fast succession, move to the positon right next to them, when they were supposed to move to an upper position (where words are formed).
Modified swapTiles and tileTapped, to add newly tapped tiles into an array. Use completion handler in saveContext to ensure that context is first saved, and therefore the letter positions are all correct, and then call swapTiles on the next tile in the array waiting to be swapped. However, maxTilesToSwap has not been seen to go over 0, which means there is never more than one tile queued up to be swapped, so this safeguard may be unnecessary.
Note that some of these  misguided tiles happen after a pan operation, where a tap was misinterpreted as a pan when the tile was moved a tiny, almost imperceptible, amount when tapping. Fix by requiring the Tap geture recognizer to fail before invoking the Pan gesture recognizer.