Create Game NSManaged object
Change the data model to include a Game object, with a one-to-one relationship with the GameData object. The properties gameName and gameID were moved from GameData to Game. A gameType property was also added to Game. Game should include properties that are set once for that type of Game. GameData should include persistent data that is unique to that user playing that game.