Fix bug where words are filled in but still findable
Upon starting a new list, some words are shown (in gray) when they should be hidden. These are words that are 'mastered', initially set to inactive, but then reset to active for playability, so there aren't too many inactive in one list. Fixed by, instead of setting all the mastered words to inactive, and then setting some back to active, set only the ones needed to be inactive. The others are left at the default of active.
Also add first unit test, test calculating the inactive quotient.