Calculate the number of bars in progress graph per device
Depending on device width, the number of bars will vary. Also, the bars are now masked with an image, rounded at top (as before), but square at bottom. Fix the issue that bars at 20 or below are a little less wide due to the rounded corners at the bottom. Bars are all the same width, and so are spaced evenly.