Implement calculateLevel in ProgressInfoViewController
Call calculateLevel from the parent view controller, which is the ProgressViewController, and display result in info box. Call calculateLevel in viewWillAppear so that it's called after the number of words in each level have been calculated. Update the calculateLevel func in ProgressViewController so that levels are calculated to the tenths place, but always round down, never up.
Also update the bar graph shadow with single line at bottom.