In ProgressView, remove empty bars on left and add on the right
If a level has 0 words, and it has no levels with > 0 words to its left, do not add the unneeded bar to the graph stack view. But so that the graph does not have too much empty space, for each bar eliminated on the left, add one on the right. The bars on the right may, or may not, have words in them.