Implement Core Data lightweight migration
Add versioning to Core Data model. Add options to persistent store coordinator to allow for inferring migration from older versions of the data model. Add a property to a Core Data object. The property added is masteryLevelManual: Bool, which is false by default. (Will allow the user to set the mastery level instead of being set automatically.) Test on simulator and on device to check whether the new data model version can be used without a crash and 'incompatible store' error. Works.