Update layout of Progress View
Move graph to top of view, with info box below. Looks better because the flat bottom of the graph now matches the flat top of the info box, whereas before, the top of the graph, which is uneven, had the info box perched above it.
Move the Done button from top to the bottom right. Add a button for leaderboards, in preparation for implementing Game Center.