Add leaderboard for Level
In iTunes Connect, add a leaderboard for the player level, formated to 1 decimal place. This will accept an int, which needs to be multiplied by 10, to account for the 1 decimal place that will be added in Game Center.
Implement saving scores to this level leaderboard. Change calculateLevel() to return a float instead of a string. Do the string conversion to display the level in-app in ProgressInfoViewController.