Add NSFetchedResultsControllerDelegate to AtoZModel
Begin to move calculation of level to model, for greater flexibility of when level is sent to Game Center. NSFetchedResultsControllerDelegate adds requirements to fulfill protocol. Requirements met by inheriting from NSObject. But NSObject's requirement to call super.init all properties to be init'd first. Hence, many changes to initialization of properties in AtoZModel.