Change InfoViewController to modal presentation style
Present infoViewController with modal presentation style .overCurrentContext, instead of as a popover. Presenting as popover was unnecessary and had limitations.
Now that the game launches with some basic info (Title and briefly how to play) in the info VC, set the initial view controller to be the main playing view. The info view controller is dismissed with a button.