Fix shadows for blurred bubble views
Move code for shadow to ViewDidAppear so that it can correctly reference the final size of views loaded from xib, to create the shadow mask. Set shadow layer opacity to 0 initially, because the shadow mask is not in place yet at start of ViewDidAppear, and the parts to be masked still show. Animate in the shadow layer opacity property.